Refactor main method signature; update sum method return type to long; add ShortCircuitDemo class

+++ src/main/java/com/infosupport/PalindromeChecker.java
@@ -54,7 +54,7 @@ public class PalindromeChecker {
         return result;
     }
 
-    public static int somArr(int[] nummers) {
+    public static long somArr(int[] nummers) {
         return stream(nummers).sum();
     }
 }
